% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{drop_edges}
\alias{drop_edges}
\title{Drop perimeter wells from plate‑based data}
\usage{
drop_edges(
  .data,
  apply_to_plates = NULL,
  drop = TRUE,
  col_edges = c(1, 24),
  row_edges = c("A", "P")
)
}
\arguments{
\item{.data}{Data frame containing `plate`, `row`, and `col` columns.}

\item{apply_to_plates}{Character vector of plates to trim.
`NULL` = all plates.}

\item{drop}{Logical; if `FALSE`, returns `.data` untouched.}

\item{col_edges}{Integer vector of columns to drop (e.g. `c(1, 24)`).}

\item{row_edges}{Character vector of rows to drop (e.g. `c("A","P")`).}
}
\description{
Drop perimeter wells from plate‑based data
}
