% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mean_sd}
\alias{mean_sd}
\title{Mean ± SD summary helper for *ggplot2::stat_summary()*}
\usage{
mean_sd(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A data frame with columns `y`, `ymin`, `ymax`.
}
\description{
Mean ± SD summary helper for *ggplot2::stat_summary()*
}
